+++ app/pages/[[locale]]/app/chat/[username].vue
@@ -7,9 +7,24 @@ definePageMeta({
 
 const route = useRoute()
 const username = computed(() => route.params.username as string)
+const { t } = useI18nExperimental()
 
 const receiverUsernameNotFound = useReceiverUsernameNotFound()
 const characterQuery = useCharacterQuery(username)
+
+useSeoMeta({
+  // HTML tags
+  title: () => t('Talk to "{name}" thal. Learn English.', { name: characterQuery.data.value?.name }),
+  description: () => t('Get translations, corrections, and listening while chatting. Try it for free to level up your English.'),
+
+  // Facebook tags
+  ogTitle: () => t('Talk to "{name}" thal. Learn English.', { name: characterQuery.data.value?.name }),
+  ogDescription: () => t('Get translations, corrections, and listening while chatting. Try it for free to level up your English.'),
+
+  // Twitter tags
+  twitterTitle: () => t('Talk to "{name}" thal. Learn English.', { name: characterQuery.data.value?.name }),
+  twitterDescription: () => t('Get translations, corrections, and listening while chatting. Try it for free to level up your English.'),
+})
 </script>
 
 <template>